python: split casadi_functions file
